move all permission related funtion to common file
